世界坐标

+++ scenes/space_rock/SpaceRockJsonReader.cs
@@ -14,7 +14,7 @@ namespace CosmicMiningCompany.scripts.data
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Health { get; set; }
+        public int Health { get; set; }
         public string Loot { get; set; }
         public int Distance { get; set; }
         public int Probability { get; set; }
@@ -44,7 +44,7 @@ namespace CosmicMiningCompany.scripts.data
                 {
                     Id = ConvertToInt(GetObjectValue(item, "编号ID")),
                     Name = ConvertToString(GetObjectValue(item, "名称")),
-                    Health = ConvertToFloat(GetObjectValue(item, "血量")),
+                    Health = (int)ConvertToFloat(GetObjectValue(item, "血量")),
                     Loot = ConvertToString(GetObjectValue(item, "掉落物")),
                     Distance = ConvertToInt(GetObjectValue(item, "刷新距离")),
                     Probability = ConvertToInt(GetObjectValue(item, "刷新概率系数")),
